MeterDetail.vue is separated to components (basic, details, readings,transactions) for code legibility
